requirements.txt
# Core Dependencies
google-generativeai>=0.8.5
python-dotenv>=1.1.1
requests>=2.32.4

# Async Support
aiohttp>=3.9.0
aiofiles>=23.2.1

# Testing Dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Code Quality
black>=24.0.0
pylint>=3.0.0
mypy>=1.8.0

# Monitoring & Logging
structlog>=24.1.0
prometheus-client>=0.19.0

# Security
cryptography>=42.0.0

# Performance
lz4>=4.3.0

# Development Tools
ipython>=8.20.0